refactored the code and created the following components: searchFeature, ThemeToggleButton, LanguageSwitcher. This also fixed the mobile navigation issue I had when clicking on ToggleTheme and flags
